+++ src/components/page-components/Resume/ResumeAutobiographyForm.js
@@ -1,40 +1,22 @@
 import React, { Component } from 'react';
-import { Input, Form, Button, Space, Upload, message } from 'antd';
-import { MinusCircleOutlined, LinkOutlined, PlusOutlined, InboxOutlined } from '@ant-design/icons';
+import { Input, Form, Button, Space } from 'antd';
+import { MinusCircleOutlined, PlusOutlined } from '@ant-design/icons';
 import './Resume.css';
 import ResumeFormTop from './ResumeFormTop';
 import { Link } from 'react-router-dom';
 
-const { Dragger } = Upload;
-
-const props = {
-  name: 'file',
-  multiple: true,
-  action: 'https://www.mocky.io/v2/5cc8019d300000980a055e76',
-  onChange(info) {
-    const { status } = info.file;
-    if (status !== 'uploading') {
-      console.log(info.file, info.fileList);
-    }
-    if (status === 'done') {
-      message.success(`${info.file.name} file uploaded successfully.`);
-    } else if (status === 'error') {
-      message.error(`${info.file.name} file upload failed.`);
-    }
-  },
-};
-
-class ResumeForm5 extends Component {
+class ResumeAutobiographyForm extends Component {
   handleChange = (value) => {
     console.log(`selected ${value}`);
   };
 
   render() {
+    const { TextArea } = Input;
     return (
       <>
         <div>
           <ResumeFormTop />
-          <h2>포트폴리오</h2>
+          <h2>자기소개서</h2>
           <Form name="dynamic_form_nest_item" onFinish={this.onFinish} autoComplete="off" className="EducationForm">
             <Form.List name="users">
               {(fields, { add, remove }) => {
@@ -47,20 +29,13 @@ class ResumeForm5 extends Component {
                         align="start"
                       >
                         <Form.Item>
-                          <Input
-                            style={{ width: 670, height: 60 }}
-                            placeholder="링크(http://)"
-                            prefix={<LinkOutlined className="site-form-item-icon" />}
-                          />
+                          <Input placeholder="제목" style={{ width: 880, height: 60 }} />
                         </Form.Item>
-                        <Dragger {...props} style={{ width: 670 }}>
-                          <p className="ant-upload-drag-icon">
-                            <InboxOutlined />
-                          </p>
-                          <p className="ant-upload-text">Click or drag file to this area to upload</p>
-                        </Dragger>
+                        <Form.Item>
+                          <TextArea rows={4} placeholder="상세내용" style={{ width: 880 }} />
+                        </Form.Item>
                         <MinusCircleOutlined
-                          style={{ fontSize: 20, marginTop: 65 }}
+                          style={{ fontSize: 20, marginTop: 40 }}
                           onClick={() => {
                             remove(field.name);
                           }}
@@ -86,16 +61,16 @@ class ResumeForm5 extends Component {
 
             <Form.Item>
               <Button type="primary" htmlType="submit" style={{ width: 150, height: 40 }}>
-                포트폴리오저장
+                자기소개서저장
               </Button>
             </Form.Item>
           </Form>
         </div>
 
         <div style={{ float: 'right' }}>
-          <Link to="/resumeAutobiography">
-            <Button type="primary" htmlType="button" style={{ width: 150, height: 40, marginBottom: 55 }}>
-              다음
+          <Link to="/resumeResult">
+            <Button type="primary" htmlType="button" style={{ width: 150, height: 40, marginBottom: 50 }}>
+              완료
             </Button>
           </Link>
         </div>
@@ -104,4 +79,4 @@ class ResumeForm5 extends Component {
   }
 }
 
-export default ResumeForm5;
+export default ResumeAutobiographyForm;
