+++ src/pages/ResumeForm.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
-import ResumeForm1 from './../components/page-components/Resume/ResumeForm1';
+import ResumePrivacyForm from '../components/page-components/Resume/ResumePrivacyForm';
 
 export default class ResumeForm extends Component {
   render() {
-    return <ResumeForm1 />;
+    return <ResumePrivacyForm />;
   }
 }
