+++ src/routes.js
@@ -6,11 +6,11 @@ import Index from './pages/Index';
 import Login from './pages/Login.js';
 import UserSignUp from './pages/SignUpUser.js';
 import ResumeForm from './pages/ResumeForm';
-import ResumeForm2 from './components/page-components/Resume/ResumeForm2';
-import ResumeForm3 from './components/page-components/Resume/ResumeForm3';
-import ResumeForm4 from './components/page-components/Resume/ResumeForm4';
-import ResumeForm5 from './components/page-components/Resume/ResumeForm5';
-import ResumeForm6 from './components/page-components/Resume/ResumeForm6';
+import ResumeCareerForm from './components/page-components/Resume/ResumeCareerForm';
+import ResumeCertificateForm from './components/page-components/Resume/ResumeCertificateForm';
+import ResumeLanguageAbility from './components/page-components/Resume/ResumeLanguageAbility';
+import ResumePortfolioForm from './components/page-components/Resume/ResumePortfolioForm';
+import ResumeAutobiographyForm from './components/page-components/Resume/ResumeAutobiographyForm';
 import ResumeResult from './components/page-components/Resume/ResumeResult';
 
 export default (
@@ -20,11 +20,11 @@ export default (
     <PrivateRoutes exact path="/" component={Index} />
     <PrivateRoutes exact path="/resume" component={ResumeForm} />
     <PrivateRoutes exact path="/resumePrivacy" component={ResumeForm} />
-    <PrivateRoutes exact path="/resumeCareer" component={ResumeForm2} />
-    <PrivateRoutes exact path="/resumeCertificate" component={ResumeForm3} />
-    <PrivateRoutes exact path="/resumeLanguageAbility" component={ResumeForm4} />
-    <PrivateRoutes exact path="/resumePortfolio" component={ResumeForm5} />
-    <PrivateRoutes exact path="/resumeAutobiography" component={ResumeForm6} />
+    <PrivateRoutes exact path="/resumeCareer" component={ResumeCareerForm} />
+    <PrivateRoutes exact path="/resumeCertificate" component={ResumeCertificateForm} />
+    <PrivateRoutes exact path="/resumeLanguageAbility" component={ResumeLanguageAbility} />
+    <PrivateRoutes exact path="/resumePortfolio" component={ResumePortfolioForm} />
+    <PrivateRoutes exact path="/resumeAutobiography" component={ResumeAutobiographyForm} />
     <PrivateRoutes exact path="/resumeResult" component={ResumeResult} />
   </Switch>
 );
